Corrected CSS

+++ FrontEnd/src/components/UserProfile.js
@@ -13,24 +13,12 @@ function UserProfile() {
         <div className="mt-5">
             <div className="container pt-5">
 
-                <Tabs defaultActiveKey="1" centered>
+                <Tabs defaultActiveKey="2 " size="large" centered>
                     <TabPane tab="Profile" key="1" >
                         <MyProfile />
                     </TabPane>
                     <TabPane tab="Room Bookings" key="2">
                         <UserBookings />
-                    </TabPane>
-                    <TabPane tab="Special Occassion Bookings" key="3">
-                        <div class="container" style = {{marginBottom : "600px"}}>
-                            <div class="row row justify-content-center">
-                                <div class="col align-self-center">
-                                <h3>Special Occassion Bookings</h3>
-                                <button className="btn btn-dark btn-lg btn-block" onClick={() => { window.location.href = "http://localhost:3000/specialoccasion/OccasionHome/ViewOccasion" }}>View All Your Occasions!</button>
-                                </div>
-                            </div>
-                        </div>
-                       
-
                     </TabPane>
                 </Tabs>
             </div>
@@ -66,13 +54,30 @@ export function UserBookings() {
     function cancelBooking(id) {
         // console.log(id)
         // console.log(bookings)
-        axios.delete(`http://localhost:8070/api/booking/delete/${id}`).then(() => Swal.fire("Success", "Booking Cancelled", "success")).then(() => { window.location.href = "http://localhost:3000/user/userprofile" });
 
-    }
+        Swal.fire({
+            title: "Are you sure you want to cancel this booking?",
+            text: "This action cannot be undone",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#d9534f",
+            cancelButtonColor: "#292b2c",
+            confirmButtonText: "Yes, cancel booking!",
+          }).then((result) => {
+            if (result.isConfirmed) {
+                axios.delete(`http://localhost:8070/api/booking/delete/${id}`).then(() => Swal.fire("Success", "Booking Cancelled", "success")).then(() => { window.location.href = "http://localhost:3000/user/userprofile" });
+            }
+          });
+        }
+     
 
     return (
         <div>
-
+            
+            <div>
+                <h3  align = "center">Hello, {user.cusname}!</h3>
+                <h5  align = "center">You can find all your bookings below</h5>
+            </div>
             <div className="row justify-content-center">
                 <div class="col-md-8">
 
@@ -100,7 +105,7 @@ export function UserBookings() {
                                 <p><b>Booking status : </b>{booking.status == "cancelled" ? <Tag color="orange">cancelled</Tag> : <Tag color="cyan">Success</Tag>}</p>
 
                                 {booking.status == "cancelled" ? "" : <div className="text-right">
-                                    <button className="btn btn-dark" onClick={() => { cancelBooking(booking._id) }}>Cancel</button>
+                                    <button className="btn btn-dark btn-block" onClick={() => { cancelBooking(booking._id) }}>Cancel Booking</button>
                                 </div>}
 
                                 </div>
